add admin and unknown endpoint middlewares

+++ back/src/routes/index.ts
@@ -1,6 +1,8 @@
 import express, { Request, Response } from 'express';
 import { IItem } from '../common/interfaces';
+import { isAdmin } from '../middlewares/isAdmin';
 import { Productos } from '../services/producto';
+import { throwUnauthorizedError } from '../utils/others';
 
 const routes = express.Router();
 
@@ -39,39 +41,51 @@ routes.get(
 );
 
 routes.post(
-  '/productos/guardar',
+  '/productos/guardar', isAdmin,
   async (req: Request, res: Response): Promise<void> => {
     try {
-      const producto = req.body;
-      const newProducto: IItem = await saveProducto(producto);
-      res.json({ data: newProducto });
+      if (req.admin) {
+        const producto = req.body;
+        const newProducto: IItem = await saveProducto(producto);
+        res.json({ data: newProducto });
+      } else {
+        throwUnauthorizedError(req);
+      }
     } catch (e) {
-      res.status(400).json({ error: e.error, message: e.message });
+      res.status(400).json({ error: e.error, message: e.message, descripcion: e.descripcion });
     }
   }
 );
 
 routes.put(
-  '/productos/actualizar/:id',
+  '/productos/actualizar/:id', isAdmin,
   async (req: Request, res: Response): Promise<void> => {
     try {
-      const producto = await updateProducto(req.params.id, req.body);
-      res.json({ data: producto });
+      if (req.admin) {
+        const producto = await updateProducto(req.params.id, req.body);
+        res.json({ data: producto });
+      } else {
+        throwUnauthorizedError(req);
+      }
     } catch (e) {
-      res.status(404).json({ error: e.error, message: e.message });
+      res.status(404).json({ error: e.error, message: e.message, descripcion: e.descripcion });
     }
   }
 );
 
 routes.delete(
-  '/productos/borrar/:id',
+  '/productos/borrar/:id', isAdmin,
   async (req: Request, res: Response): Promise<void> => {
     try {
-      await deleteProducto(req.params.id);
-      const productos = await getProductos();
-      res.json({ data: productos });
+      if (req.admin) {
+        await deleteProducto(req.params.id);
+        const productos = await getProductos();
+        res.json({ data: productos });
+      } else {
+        throwUnauthorizedError(req);
+      }
     } catch (e) {
-      res.status(404).json({ error: e.error, message: e.message });
+      res.status(404).json({ error: e.error, message: e.message, descripcion: e.descripcion });
     }
   }
 );
